# JeninaLanziU2M2Summative/stwitter-service/stwitter-openapi.yaml
openapi: 3.0.0
info:
  title: Stwitter Service
  description: A stripped-down social media service where users can post and add comments to those posts.
  version: 0.0.1
  
servers:
  - url: http://localhost:8889
    description: Local dev server
    
paths:
  /posts:
    post:
      summary: Creates a post.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: Created
        '400':
          description: Invalid request syntax.
        '422':
          description: Incomplete information. Some required fields missing.
  /posts/{id}:
    get:
      summary: Returns a post by id.
      parameters: 
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: A post.
  /posts/user/{posterName}:
    get:
      summary: Returns a list of posts by the same poster name.
      parameters: 
        - name: posterName
          in: path
          required: true
          description: The name of the poster you want to search by.
          schema:
            type: string
            example: ben
      responses:
        '200':
          description: A JSON array of posts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
  /comments:
    post:
      summary: Creates a comment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
              
components:
  schemas:
    Post:
      properties:
        postId:
          type: integer
        postDate:
          type: string
        posterName:
          type: string
        post:
          type: string
      required: 
        - postDate
        - posterName
        - post
    Comment:
      properties:
        commentId:
          type: integer
        postId:
          type: integer
        createDate:
          type: string
        commenterName:
          type: string
        comment:
          type: string
      required: 
        - postId
        - createDate
        - commenterName
        - comment